refactor: separate presets and models into their own folders
